refactor(store): exact money (Decimal) and typed scope enum
- Money fields (quotas.quota_total/cost_used, usages.cost,
  usage_rollups.cost) move from f64 to rust_decimal::Decimal, stored as
  exact decimal TEXT — lossless and portable across sqlite/pg/mysql/file
  (no sea-orm numeric feature). Records serialize via serde-with-str so
  the file backend's JSON stays exact; the affected db entities regain Eq.
- scope String -> enum Scope { Org, Team, User } (records, re-exported);
  route_permissions / rate_limits / quotas and the trait / impls / libsql
  stubs / identity cascade callers take Scope; stored as text via
  as_str()/parse.

Adds rust_decimal (shared dep). New regression test asserts a Decimal
quota survives the sqlite round-trip with exact equality. 19 tests pass;
build/clippy --features full, default build, and wasm32 --features edge
all green.

+++ src/store/persistence/records/usage.rs
@@ -1,6 +1,7 @@
 //! Usage records (§8-D): per-request usage and rollups. All tables are
 //! append/accumulate only; retention is deferred.
 
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 /// A per-request usage row (one logical proxied request, §8-D).
@@ -25,7 +26,8 @@ pub struct Usage {
     pub cache_read_tokens: i64,
     pub cache_creation_5m_tokens: i64,
     pub cache_creation_1h_tokens: i64,
-    pub cost: f64,
+    #[serde(with = "rust_decimal::serde::str")]
+    pub cost: Decimal,
     /// Unix seconds.
     pub created_at: i64,
     /// Unix seconds.
@@ -52,7 +54,8 @@ pub struct UsageInput {
     pub cache_read_tokens: i64,
     pub cache_creation_5m_tokens: i64,
     pub cache_creation_1h_tokens: i64,
-    pub cost: f64,
+    #[serde(with = "rust_decimal::serde::str")]
+    pub cost: Decimal,
 }
 
 /// An aggregated usage bucket for one `(granularity, bucket_start, dimensions)`
@@ -73,7 +76,8 @@ pub struct UsageRollup {
     pub requests: i64,
     pub input_tokens: i64,
     pub output_tokens: i64,
-    pub cost: f64,
+    #[serde(with = "rust_decimal::serde::str")]
+    pub cost: Decimal,
     /// Unix seconds.
     pub created_at: i64,
     /// Unix seconds.
@@ -95,5 +99,6 @@ pub struct UsageRollupInput {
     pub requests: i64,
     pub input_tokens: i64,
     pub output_tokens: i64,
-    pub cost: f64,
+    #[serde(with = "rust_decimal::serde::str")]
+    pub cost: Decimal,
 }
